+++ NEWS.md
@@ -1,5 +1,14 @@
 # Changelog soilptf
 
+## Version 0.4.0 2023-04-29
+
+### Added
+* predict functions for all ptfs: `ptf_cec`,`ptf_phbc`,`ptf_mwd`,`ptf_wsa`,`ptf_hwc`,`ptf_sss`,and `ptf_cdec` 
+* package table `sptf_soilproperties` updated for metals, hwc, cdec, whc and paw
+
+### Changed
+* script names for functions are updated, making them consistent with function names
+
 ## Version 0.3.0  2023-04-27
 
 ### Added